server/requirements.txt
pypresence==4.3.0
flask==2.3.3
requests==2.31.0
pillow==10.0.1
python-dotenv==1.0.0